requirements.txt
cryptography==41.0.7
Flask==3.0.2
imageio==2.34.0
moviepy @ git+https://github.com/Zulko/moviepy
Pillow
pydub
PyJWT
PyMySQL==1.1.0
